Fix keyboard visibility issue
