+++ app/src/Components/Common/KeyboardAvoidingViewScroll.tsx
@@ -1,12 +1,12 @@
 import React, { PropsWithChildren, useContext, useEffect, useRef } from 'react';
 import { ScrollView, StyleSheet } from 'react-native';
-import context from '../../Context/context';
 import { KeyboardAvoidingScroll } from '../../Helpers/StylizedComponents';
-import { AppContext } from '../../types/AppContextType';
+import authContext from '../../Context/authContext';
+import { AuthContextType } from '../../types/AuthContextType';
 
 const KeyboardAvoidingViewScroll = ({ children }: PropsWithChildren) => {
   const scrollViewRef = useRef<ScrollView>(null);
-  const { isKeyboardVisible } = useContext(context) as AppContext;
+  const { isKeyboardVisible } = useContext(authContext) as AuthContextType;
 
   useEffect(() => {
     setTimeout(() => {
